feat: replace restaurant home page with immersive layout
